refactor: ApiError 타입 가드 api 폴더로 이동
