+++ frontend/src/features/auth/types/api.ts
@@ -30,17 +30,3 @@ export type ApiError = {
     code: string;
     message: string;
 };
-
-/**
- * 다양한 에러를 안전하게 ApiError 타입으로 변환
- */
-export function toSafeApiError(error: unknown): ApiError {
-    if (error && typeof error === "object" && "status" in error && "code" in error && "message" in error) {
-        return error as ApiError;
-    }
-    return {
-        status: 500,
-        code: "UNKNOWN_ERROR",
-        message: error instanceof Error ? error.message : "알 수 없는 오류가 발생했습니다.",
-    };
-}
